# crimson/net: implement preemptive shutdown/close