--- a/src/crimson/net/Socket.h
+++ b/src/crimson/net/Socket.h
@@ -22,6 +22,11 @@ class Socket
   seastar::input_stream<char> in;
   seastar::output_stream<char> out;
 
+#ifndef NDEBUG
+  bool down = false;
+  bool closed = false;
+#endif
+
   /// buffer state for read()
   struct {
     bufferlist buffer;
@@ -38,6 +43,12 @@ class Socket
       // the default buffer size 8192 is too small that may impact our write
       // performance. see seastar::net::connected_socket::output()
       out(socket.output(65536)) {}
+
+  ~Socket() {
+#ifndef NDEBUG
+    assert(closed);
+#endif
+  }
 
   Socket(Socket&& o) = delete;
 
@@ -79,12 +90,20 @@ class Socket
     return out.write(std::move(buf)).then([this] { return out.flush(); });
   }
 
+  // preemptively disable further reads or writes, can only be shutdown once.
+  void shutdown();
+
   /// Socket can only be closed once.
-  seastar::future<> close() {
-    return seastar::smp::submit_to(sid, [this] {
-        return seastar::when_all(
-          in.close(), out.close()).discard_result();
-      });
+  seastar::future<> close();
+
+  // shutdown input_stream only, for tests
+  void force_shutdown_in() {
+    socket.shutdown_input();
+  }
+
+  // shutdown output_stream only, for tests
+  void force_shutdown_out() {
+    socket.shutdown_output();
   }
 };