After this change, the messenger should be able to:
* preemptive shutdown `Socket`
* preemptive shutdown `Connection`
* preemptive shutdown `Messenger`

No impact on the I/O path.
Also added unit-tests for messenger.

Depends on #28623