@cyx1231st 2019-06-24
I think there's no performance impact because the changes are not at the IO path, here's the perf result:

* before:
```
$ perf_crimson_msgr
perf settings:
  client[>> v1:0.0.0.0:9010/0](bs=4096, ramptime=5, msgtime=15, jobs=1, depth=512)
  server[v1:0.0.0.0:9010/0](bs=0, core=0)

    sec depth    IOPS    MB/s lat(ms)
1.00202   456  223629 873.549  2.0329
  1.002   399  223948 874.797 2.03166
1.00203   512  226276 883.889 2.00665
1.00243   512  223721 873.909 2.03191
1.00203   416  222743  870.09 2.04286
1.00201   389  223072 871.377 2.04216
1.00198   434  223778 874.133 2.03273
1.00302   387  222842 870.476 2.04092
1.00197   512  224053 875.209 2.03218
1.00253   467  219177  856.16 2.07971
-------------- summary --------------
10.0235     -  223310 872.305 2.03726

1.00349   326  220496 861.311 2.06584
1.00297   512  223913  874.66 2.02636
1.00199   512  222942 870.869 2.04141
1.00203   415  224452 876.764 2.02613
1.00499   370  224151 875.588 2.02731
-------------- summary --------------
  15.04     -  223266 872.133 2.03731
```
* with the PR series:
```
$ perf_crimson_msgr
perf settings:
  client[>> v1:0.0.0.0:9010/0](bs=4096, ramptime=5, msgtime=15, jobs=1, depth=512)
  server[v1:0.0.0.0:9010/0](bs=0, core=0)

    sec depth    IOPS    MB/s lat(ms)
1.00175   493  222641 869.691  2.0445
1.00248   328  223287 872.215 2.03818
1.00222   452  223807 874.245 2.02927
1.00148   436  223161 871.723 2.03692
1.00251   469  223895 874.588 2.02945
1.00251   397  223175 871.777 2.04125
1.00248   459  219653 858.019 2.07922
1.00201   512  223062 871.335 2.04199
1.00196   512  222515   869.2 2.04544
1.00204   491  219680 858.124 2.07762
-------------- summary --------------
10.0229     -  222468 869.017 2.04629

1.00249   367  222166 867.834 2.05017
1.00197   512  223464 872.906 2.03983
1.00254   332  223004 871.108 2.04281
1.00199   461  223086 871.428  2.0385
1.00253   437  223753 874.035 2.03447
-------------- summary --------------
 15.035     -  222678 869.837  2.0446
```